Ensure we return an error when creating a stream
Additionally, we need a path to already exist that we're attempting to
watch on. For now /var/lib/juju/locks, we can always move this in the
future, we just need to ensure the folder exists already.

This is all nicely wired up and you can experiment blocking the change
stream. If we're then having problems with the db becoming out of
control, inserting one file in the right location will block the change
stream and cause it to behave itself.