Force ldflags on go test
For some strange reason, you sometimes need to force the ldflags on go
test. It's not always required, but only sometimes and normally it's
required on jenkins.

This is very strange and can be machine dependent, so clearly there is
some sort of missing link somewhere when setting stuff up, that requires
it to be set.