EventQueue change stream
The following changes the change stream from just returning a queue,
instead it returns an event queue. An event queue can multiplex out
changes to multiple sources.

The basic idea is to create a subscription model that all the watchers
can subscribe to. They can then peel off a change and they can either
block the channel to ensure consistency or send push that into their own
channel to perform it async.

There is a lot refactoring going on here, as we're building more and
more functionality.

This really needs to be battle tested as this is the core of the new
Juju system.