Use file based sqlite instead of memory mode
Unfortunately we can't use memory mode for our sqlite tests, as it's
possible to create a race in sqlite where the data isn't correct and
we're seeing old data. The solution is to create a new file for every
test and use that for our testing purposes.

I've ensured that we also close the db unlike before and that we've
added some additional logs to help with test failures.