Add a test for file blocking
The following adds a test for file blocking, this shows how file
blocking can prevent the receiving of a notification. Unfortunately, we
have to wait for a second to ensure we don't see a change, which is
annoying as it blows out our test suite times. I don't think there is a
nice way to do this, unless we know exactly when a change is triggered.