Update dqlite build shas
The dqlite build shas have been rebuilt to update to the latest dqlite
and raft implementations. See[1]

This PR ensures that once they've been built we can then download them
and build them.

You can verify the build passed with all the correct shas at the jenkins
build page[2]

As a driveby I've now remove the fallback, which attempted to just use
the latest dqlite library if the sha wasn't found. This no longer used
and is deprecated.

Lastly we still need to find a way to ensure what we built is valid
based on the sha or the BOM? That can happen in a subsequent PR.

 1. https://github.com/juju/juju/pull/15208
 2. https://jenkins.juju.canonical.com/job/build-dqlite/55/