Add refresh resource tests.
The tests in refresh_resources_test.go were flawed, replacing with new tests in the
refresh_test file. The old file had tests which contacted the charm repository and
had issues with mongo. The new tests do not.

The majority of the refresh with resources test logic is in utils. The tests in
RefreshSuite check basic expected results.