Back firewall rules commands with model config
Add model-config items
- ssh-allowlist
- application-offer-ingress-allowlist

Use model-config internally instead of the firewall rules endpoints.

This has the side effect that the default value for
juju-application-offer firewall rule of "0.0.0.0/0" instead of returning
a NotFound error.

Remove firewaller falling back to public is juju-application-offer is
empty. Since the it can only be empty if explicitly set so

Deprecate firewall-rule commands

The firewall rules api has existed for a while, not being implemented.
Now that it needs to be implemented, storing this kind of model config
together in model-config makes more sense. Also, ssh and
juju-application-offer both behave quite differently, so doesn't make
sense to keep them together