Ensure we close the subscription channel
The following ensures that the subscription channel is closed when the
unscription happens. This happens asynchronously because closing a
channel during a dispatch can cause a panic. We can guarantee the
subscription will not be called in another dispatch because the removal
of the subscription is done synchronously. Defering the closing to the
next tick with in the loop. So you'll always guarantee to see the event
you're trying to unsubscribe from in the same dispatch loop.