Tracked DB Type
The following wraps the *sql.DB type to enable the tracking of the db
from other locations. This is useful to know when a db has been put into
an error condition (tainted). By exposing an Err method on the tracked
db type, then allows consumers to check if the db is useable or not. We
might not need to expose this everywhere, but for low level workers this
becomes extremely useful.

In addition to this, we can the poll the db at a given interval to check
that the db is accessible. We don't need to do super frequently, but
enough to validate that it's working.
From there we can then bring down individual DBs where required or crash
the whole dbaccessor.

The code changes are rather invasive as we're now pushing the tracked db
type everywhere. I'm just glad we're doing it now and not later down the
stream.

Additionally, I'm unsure we should offer out the whole *sql.DB, but
instead offer out just the eventqueue (in PR#15204) and a Txn method, so
correct encapsulate the retry and commit/rollback semantics.