Use callback instead of channels
Channels become really hard to use correctly when attempting to close.
There are potential races and pitfalls. To clean the model up, we'll
move to callbacks. This cleans up how the event queue will work. It's
much nicer to use and a lot more simpler.

Additionally, there are less footguns, which for a complex application,
this is a requirement.