TrackedDB now closed over a db and txn
The following changes the whole semantics of a tracked db type. It now
closed over a db and txn. This is to allow the realisation of a db at
the very last second. At no point is a db allowed to be accessed in a
raw manor. By doing this, we can swap out the underlying db if there is
a detectable fault.

The side effect of changing this is that we now need to rework a lot of
the code to correctly handle the changes. Leases can no longer prepare
statements easily, as a prepared statement is tracked with a db. Closing
that db means that prepared statement is no longer valid. We'll have to
potentially come up with different semantics for that to work, but for
now we can just rely on the fact that transactions are fast. Potentially
we could have a Register that will be invoked everytime a db is swapped
out and the statements could be re-registered again.

Currently, retry hasn't been implemented, but that will be done next.