Serialize all operations in the event queue
This allows all operations to be correctly serialized including the
dispatching signalling logic.

Unfortunately, this requires all subscriptions to become tombs and the
event queue to also become a catacomb, this allows everyone to know when
a subscription is done. We tie a tomb to a catacomb lifecycle.

This makes things a lot more complicated internally, but it is a lot
simpler for the consumer.