Add the scaffolding types and interfaces for different types of deploy
This introduces DeployerKind, which is an interface with a method
named Read to return a deployer type. The idea is that all the
different types of deploy wrt what's being deployed (e.g. local charm,
repo bundle, etc) will implement this interface and in the GetDeployer
we just call Read. This consolidates the logic that determines the
type of deployment into GetDeployer function without introducing type
specific checks etc (that are baked into the corresponding Read
methods).