Move back to channel multiplexing
It was desided that everyone hates callbacks. Although callbacks allowed
us to move to a better setup, the benefits didn't out way the cost.
Instead we're back to channels, and we can have all the goodies thanks
to moving the subscription to a tomb.

There was some thought about putting an error as the type of the done
channel, but I'm actually unsure what that gives you. We don't have
anyway of changing the error type for a subscription tomb. So it'll end
up just being tomb.ErrDying which isn't helpful. We can always revisit
this at a later date.

Interestingly, we now have subscriptions and unsubscriptions that are
synchronous, which is what we wanted all a long. The tests pass nicely
as well.