Set model ingress using ssh-allowlist cfg item
If an environ satisfies a new ModelFirewaller interface, use in the
firewaller and common/bootstrap to configure model firewall, instead of
trusting the provider to handle this for you

At the moment, this is only implemented for aws

To achieve this, a new endpoint was added to the firewaller API to
determine whether a model is a controller model. This allows us to treat
the controller firewall differently to others.

This is a compromise. Previously, the api-port and the like was opened
universally. This is an improvement, but ideally the controller
application should open it's api ports, not the model, so this branching
can be removed

NOTE: Non-controller models no longer open the api-port. The only reason
they did in the first place was a hangover from old code