Move ambiguous path check up fronte
So if there's ambiguity in the provided path, like if it's not a valid
charm/bundle path but it does exist on the fs, then we gotta catch it
before the "is this local" check otherwise in the code path it falls
into the repository charm/bundle branch which is really not ideal. So
this moves that check up front for that reason.