Prevent panic by not closing the changes channel
There is a logical race between the closing of the tomb and closing of
the active channel. By closing the catacomb and the tomb, it forces the
closing of the loop, which then closes the changes channel. At the same
time we're attempting to signal a change, which causes a panic.

The solution is to not close the changes channel, as we already have a
done channel, the consumer should just use that. That reduces the
complexity of the subscription.