Add a note about tracking model list for GetDB
To prevent any dbs being opened by accident, we should verify the db
wanting to be accessed is valid in the controller list model. We can do
that once we've moved the model list to the controller db. For now we
can leave the TODO explaining that this is required in the future.