Remove bespoke dqlite REPL
The bespoke dqlite REPL was added a long time ago, way before there was
a dqlite shell. The following replaces the custom REPL along with the
custom socket.

Instead we can now switch over to the offical dqlite shell. This means
we get to take advantage of all the latest security updates and hot
fixes.

We don't install anything by default, as we don't want to leak any
security holes to potential intruders. Instead we have a make script
that will allow you to ssh into a machine, install some dependencies and
then run another command to attach to the dqlite shell. We don't do this
in one command for two reasons:

 1. The dqlite shell doesn't correctly work over ssh with interactive
tty. The liner package doesn't know when a line terminates correctly.
 2. We can now paste a VERY big warning about you're entering dangerous
territory attemtpting to edit a live database.

Overall this reduces the code that we have to maintain and reduces any
potential vulnerabillities that may occur.

In the future it might be wise to uninstall the dqlite shell once the
process is done. For now we can leave as is.