Handle TODO changing values in DeployInfo.
There is no longer a need for the charm url to be returned to the CLI. Especially as we
intend to replace it with a identifier. Return the name and revision of what was deployed
instead. The remaining bits of what would be in the charm url are already contained in
the output.