Introspection format debug argument
The following allows the introspection tools to be provided with an
additional formatting argument. Currently the debug format output was
choosen (debug=1) as you didn't need any additional tools to read the
output from the output of the tools. Except this means that we can't
always take advantage of tooling that helps track down leaks or stopped
goroutines. The documentation states that "Passing debug=1
writes the legacy text format"[1]. Unfortunately, a lot of tooling takes
advantage of debug=2.

The solution is simple, have a default of 1, but allow passing in
another option. For example:

    juju_goroutines 2

This is backwards compatible with what tooling we already have.

 1. https://pkg.go.dev/runtime/pprof#Profile.WriteTo