DBAccessor metrics
The following adds db metrics to the dbaccessor worker. The metrics aim
to track the following metrics:

 1. How long a db request takes, along with if the result was a success
or a failure.
 2. How many concurrent db requests we get at one moment in time. Hence
the use of a guage over a counter. If we want to know how many requests
in total, we can use the number of durations from 1.
 3. How many concurrent txns we have. This is similar to 2, as they both
tend to lead from each other, but more importantly I'm looking to know
how many txns we've actually had to do. This unfortunately, won't tell
us which query we had to reply the most, but can be done via the
go-dqlite PR[1a].

Interesting to note, we don't need the duration of the total txn (all
retries), as that's the duration of the DB metric. There won't be much
information to gain from that. I also didn't include the duration for
one txn, as I think that will be too expensive with potentially not much
gain.

The code is simple, although I had to use a returned name error argument
in order to work out if it's a success or an error.

  1a. https://github.com/canonical/go-dqlite/pull/227