Download charm on suggestion
The following will attempt to download a charm of the suggested output
if there is one given. This is a simple crude implementation that looks
at the return from the suggested output before attempting another go.

This is really handy if a charm hasn't got the latest LTS support.

The code just recursively calls refresh again if there is a suggestion.
It lets the user know that the original series supplied wasn't supported
and is going to download the next series from the response. The code
ensures that we try the next one from the result set in an ordered
manor. We don't use the sorted by name for this occasion.

This should remove some of the pain points of juju download for users.