Cloud schema
The following adds the cloud schema to the controller DDL. This is
pretty much a 1-1 replica of the existing cloud doc from the state
package. The only difference is we've normalized all the data so that we
know with certainty what a type is for a given release. For example, we
know all the cloud types that we support as a hole, but we might not
actually support each one at runtime. For example, we have a light
weight k8s version that doesn't include the normal cloud types like
azure, ec2, gce et al.

Additionally, I've also included all the cloud auth types that we could
possibly support, yet each provider might only support a subset of them.

Doing it this way ensures that there are known quantified data that we
model. It prevents each type becoming a data bag of potential string
variations and limiting mistakes. Modeling this should then just
fallout.

For cloud_auth_types, cloud_regions and cloud_certificates these are
link tables, so selecting by the uuid for each cloud will return their
data.

I've not put a index on cloud_certificates as I don't believe we'll
search by certificate directly, but that might be proven wrong later on.