Refactor caasPrecheck.
To enable reuse of the content by DeployFromRespository as well as the old Deploy code.
caasDeployParams holds the data to be prechecked, as using the old Deploy params is not
viable in the new world of 2 very different Deploy methods. Precheck takes the pieces
required to do the validation as arguments.