Thread db getter trhough to the apiserver
The following pushes the db getter through the many, many, many layers
of the apiserver so that we can use it at the facade level. The work to
get the db directly hasn't been implemented, just the scaffolding for
that.

In part of doing this, I moved the DBGetter out of the worker into the
core database package, this makes it a lot cleaner when attempting to
use things from the apiserver.