Expose ControllerDB on the context
The following exposes the ControllerDB on the context, as we're only
doing controller db related changes it makes sense to type this. DQlite
doesn't currently support the removal of the db completely, so we want
to prevent the creation of a db by accident.

Also this exposes the underlying TrackedDB, not a raw *sql.DB. It is up
to the user of the tracked db to ensure that the db isn't stale. At no
point should you hangon to the db reference. A new reference should be
sought after everytime.