Fix destroy-model/destroy-controller to handle --force better.
--force and --timeout/--model-timeout now properly control
the destruction of models so that --force with --timeout
properly progresses despite the status of the model/environment.

--force without --timeout now only propagates force to the
cleanup of model entities, rather than the model removal
itself.

Since model destruction cannot be stopped once it has started,
--timeout no longer makes sense to be passed for non-forceful
model destruction.

Graceful model destroys now wait indefinitely
for the entities to cleanup and the cloud resources to cleanup
before removing the model from state. Due to this change it
is now required to be able to change the model destruction
parameters while a model destroy is being processed by the
undertaker. The undertaker will now restart itself when
ForceDestroy or Timeout changes.