Ensure we call timer reset
The trackedDBWorker was missing a timer.Reset once a verification had
occured. Luckily, a CR spotted this before it ever reached production.

The code is simple, it just calls timer.Reset at the end of the timer
loop. The tests were changed to ensure that the reset was called with
the right interval and the right number of times.