Fix endpoint parse at lxd bootstrap
When bootstrapping lxd, we detect the hardware charachteristics in order
to define if we are running a local lxd cluster and therefore set the
arch accordingly.
The DetectHardware method was incorrect since it
didn't parse the endpoint correctly, this is fixed by accepting
different types of endpoint formats, as the added tests show.