Add check for explicit base in bundles when image-id constraint is set
This check was already present for charm deploys, now we add it to the bundles deploy. We do this by iterating over the list of applications in the (merged with overlay) bundle, and if we found one using image-id constraint, then we look for the base (series) to be explicitly provided.