Fix base bundle checks on constraints
The image-id constraint is not allowed on base bundles, only on overlays. In order to enforce this, a check was performed on the base bundle using the method charm.ExtractBaseAndOverlayParts(). The issue is that this method does not return the original base bundle but the resulting merged one with overlays.

This patch removes this call and adds a new check only on the original
base bundle.