Improve test coverage for changestream
The following improves the test coverage for the change stream package.
Going back to the original code, knowing what we know around the context
requirement for all requests means that changes to the code is required.

The learnings from the lease package was that every context to a txn
needs to be tracked via a tomb. If we don't track the txn with a tied
context to a tomb, then we may end up with a deadlock. This deadlock can
not be unlocked without a restart of the service.

Tests have been added to improve convidence in the interweaving of
changes from the readChanges method.

Additionally, I've taken another stab at fixing the
TestOneChangeIsBlockedByFile which is currently an intermittent failure.