Prevent memory leak when tying to a tomb
The following cancels the context when the function is done, thus
freeing it from the tomb lifecycle.

I think we really need a better pattern here as we're going to forget
about this when we run in the API server requests or similar. Either we
ask the tomb for a context without passing one in, something like
tomb.BackgroundContext() or tomb.RootContext(), which always return a
cancel back, forcing you to deal with it. OR we need to force txn to
take a tomb, so that we can tie the context correctly to the tomb.
Either way I don't think we've got this quite right yet.