New http client for lxd connections.
A change introduced to Juju introduced the concept of re-using the HTTP
client for LXD connections to multiple different servers. The LXD client
however when passed it's own http server modifies the transport for the
http client. This causes the same http client to keep having it's
transport modified for different connections.

The result is that https validation fails for some lxd remotes because
the transport is configured for a different server.

Fixes LP2003135