LXD dependency update
The following brings in the latest lxd dependency update. This update
should alleviate any potential issues we have when trying to build a
jujud on a darwin system (macOS). This is split out from PR#15183. By
splitting this out, it's possible to see what is required for updating
just the LXD part.

The part which will unlock the build process on mac is the following
PR[1] I submitted prior to the LXD project.

This will also lock sync the dqlite dependency for both juju and lxd for
now. We'll have to find a way to untaggle that in the future.

  1. https://github.com/lxc/lxd/pull/11376