Find subnet via vpc association in manual-deploy-aws
This is because the previous method, using a Name tag, was causing
issues. For a reason I was not able to work out, the subnet was
consistantly losing it's Name tag about 6-12h after creation

So don't use the tag to identify our subnet. Instead use the id of the
associated vpc. This is probably how we should have done this initially

Also make some small logging improvements, and set -e instead of
creation functions so that failures here cause the test to fail
immidiately