Fixes data races and security bugs introduced.
In Juju 3.2 work was done to start allowing third party external
authentication providers. Specifically with the introduction of this
work it allows for authentication with Juju using JWT tokens.

Because Juju up until this worked exclusively with it's own
implementation of macaroons a new authentication flow had to be
introduced into Juju along with macaroons. The way this was implemented
had most of the generic authz and authn code within the APIServer
becoming an expert in which method was in play.

As part of the above mentioned work two significant problems were
introduced into the apiserver.
- The first a race condition that had 2 or more go routines writing to
  the same region of memory with no concurrency controls. These go
  routines were trying to decorate the global apiserver state with
  context on which entity was currently logged in for a connection.
- The second was there existed cases where the permissions of one user
  could be used as the permissions for another separate user.

This change also aims to bring some uniformity around this refactor by
putting implementation details in duck typed structs and having all of
the authz and authn code operate on the duck types.