Retry debug-log on connection closed
When the controller agent restarts the connection via the API server is
cut. This becomes inconvenient when debugging, as you now have to change
your current context to restart the debug-log. Instead the juju debug-log
should do this automatically. Or at least allow you to force a retry.

The following PR just retries the API client request if there is a specific
failure. Retrying is opt-in at the moment, if we find it really useful, we
can have it opt-out. For backwards compatibility we always unwrap any error
from the retry.Call function, so that we only ever show the underlying
error and not the retry one.