Slow query log file
The following forwards all the slow log queries to an independant log
file. This should elevate the problem when you see a lot of runtime
stack traces in the normal log messages. As the runtime stack traces
look like panics, it has a tendenency to misread the logs.

Additionally, the logs get filled up all the debug stack which makes
it harder to understand what's going on in the logs. With the side
benefit of reducing the load on the database, which is where we store
the log currently.