Ignore checks for explicit base when placement directives to inexisting machines
If an application in a bundle is deployed to a specific machine, and
this machine uses the image-id constraint, then we check that the base
is explicitly defined in the application.

This patches the case where the bundle contains a placement directive to
a machine not defined in the bundle, for example lxd:0.