Make query logger into a real worker
The following changes moves the slow query logger into a fully
functional worker:

 1. It allows us to fully control the lifecycle of the logger. That
    was harder to do, as the NodeManager had to start to understand
    how things are born and die.
 2. The NodeManager was starting to know way to much about logging.

The code is relatively the same, just a lot of scaffolding to make
it a real worker.

There is a real nasty bug with the metrics pattern of registration
and unregistering, that we'll need to iron out. If you register then
get killed before a dependency is up, then you've never unregistered
and so can never bring the worker back up.