Better default make target
This solves two problems at once.

 1. There is a vscode plugin that is stalling because the default
    make target we currently have set is to build the whole of
    dqlite in a sub lxd container. This because we include a sub
    make file, which has the dqlite build target as the first
    target. This in turn causes problems with vscode itself.
 2. From a UX perspective, we shouldn't be attempting to build
    anything from a default make target sense. As the project has
    grown there isn't anything that feels like it should be default.

Instead the default make file target is help. Makefiles don't have
the ability to explicitly set the default target, instead it just
picks the first one. So the change is to put help right at the top,
which forces it to be the default one.