Unregister metrics on error
The following tweaks is a observation when testing another PR, where
failure to start a worker would prevent the worker comming up. This is
because the metrics where already registered and we don't allow a
double add. The solution is to unregister the metrics when we hit any
error case.

An alternative solution is to always unregister before a register, but
from memory there is a sticky situation in edge cases. I don't remember
the exact reason, it was over 4 years ago, just that, we shouldn't.

There are other locations that also register, but do so inside the
worker loop. This seems like an anti-pattern,  as we're now pushing down
the prometheus registration into the worker. The worker doesn't need
to know about that and we're coupling a worker to the register. This
violates cross cutting concerns of the worker.