Allow building jujud on macos
The following installs musl on darwin via homebrew, which in turns
allows building jujud. Once musl is installed there is no change
to the make file as that's correctly setup to handle this case.

As sha256sum isn't available by default on a mac, we need to check
it's available otherwise the error is rather cryptic.