Enable ModelFirewaller for OpenStack
This work has already been done for aws. Add OpenStack to the list of
providers which support this

This was mostly trivial, and involved removing a lot of responsibilities
from the OpenStack provider

This has involved updating the go-goose dependency

Also as a flyby fix an bug in config. ssh-allow was defaulting to the
wrong value if the entry was absent. Write a test for this case as well