Expose all watcher attributes
In order to help with the dashboard UI and to remove redundant subsequent
calls to the API server for the same data, we just expose more information
from the all watcher.

The code turned out to be relatively simple, we just weren't exposing it.
Although, I'm starting to believe the all watcher is anti-pattern, by
sending data with the changes, rather than just what's changed leading to
a better query API. The downsides with the all watcher is that the data
being sent has to accomindate all the data and it very much feels like
we're passing data via a back channel. The API itself would be better
served with a better API.