Fix bootstrapping continuing in IAAS models
Bootstrapping continues even during cancelation. It's hard to
replicate because it requires timing the case just right. Although
I've personally run into this many of times whilst editing the
manifest files. So to replicate it, requires configuring a broken
manifest config and bootstrap with that.

The code just reuses the same setup as CAAS and just ensures that
the context has an error or not. This is a change from the original
which checked the done channel. This makes the function call a lot
simpler.