External controller schema
Add external controller schema. This models itself based on the
existing entities in state and from the prior work from @manadart.
The only difference is that we're also modelling the external
controller model.

I've made a note about not correctly modelling the foreign key
constraint on the controller, as we don't actually have a controller or
model table. That can be added at a later date.

Also I've added a test to ensure that the tables are present if you
add any new schema. This should ensure that the schema is at least
parsable. It doesn't check for behavior or correctness of the schema.