Remove 32-bit architecture support
The following removes 32-bit architecture support. Although we've
dropped support for 32-bit architectures, we didn't actively remove
the code to drop support. As we no longer want to support 32-bit word
sizes with dqlite, as the number of jujud permutations explodes and
so it's best to completely remove the support.

The brings in the the juju/utils arch support, with the addition of
a UnsupportedArches field, which will ensure that we can at least
say what we use to support, but no longer do.

Additionally, I've dropped support for ppc and we explicitly callout
ppc64(le|el).