Model List
Adds a very basic simple model list idea to verify that a model
namespace exists before being allowed to access the db. The code
touches a lot of the dbaccessor code, which ensures that when
attempting to open a database it's in the model_list table.

It is expected that we fill out more of the model_list table, but for
now it just shows that it is possible.

Additionally, there is a namespace cache, so that we don't have to
verify the model namespace exists for a given duration (roughly 1
minute). As the model for getting a tracked db based on a model is
now late binding, we'll need to potentially request the db from the
apiserver level and this is to prevent a thundering herd problem on
start up.

The code is actually quite logical, if the namespace is the controller
namespace, then we perform no checks, everything else if fair game.

To verify that the model manager can insert the crucial model then
we add it in the model manager. Although I think that we really need
to come up with a pattern on how to do this nicely. Mixing both the
conn suite and the dqlite suite is very painful.

Note: tests in the model manager are currently broken until I solve
the conn suite/dqlite suite problem.