Remove unsupported arches from providers
To prevent showing architectures from a provider that can't be used
by juju, we filter the ones we can never support. For example the
removal for armhf, ppc and i386 which are all 32 bit word size.

In addition, we filter them using a set collection to duplicates are
removed from the output. This leads to a much cleaner output to the
user making it much cleaner.