Add ResolveForDeploy.
It is a combination of ResolveWithPreferredChannel and GetEssentialMetadata. During resolve,
the data necessary for GetEssentialMetadata is available. Combining reduces the need for
multiple calls to charmhub. This is possible with the new server side deploy work.

If we're deploying by channel, the current resource data will also be available. As resource
revisions are tied to channels, when deploying by revision, this data is not available.
Returning current resource data is best effort.