Add image-id to ec2 provider (aws)
We add the image-id constraint to the aws provider. This is done by
replacing the ec2 ImageId provided by the config with the user-provided
value in the image-id constraint.
No checks are done on the value of the constraint, this means that the
user must provide a correct image-id value for aws.

Bonus: fix small condition check on the maas provider to use the
HasImageID method of constraints.