Use model tag filter to list all ec2 instances
Previously, we opted foer sec group attachments instead. This caused
issues because before the sec group exists we need to wait for time out
before we can conclude there are none.

Recently we changed how to tag ec2 instances so tags are created at the
same time as the instance itself, meaning we no longer need this work
around