Model List Domain
The following implements the model-list via the model manager domain.
The purpose of this set of changes to ensure that we restrict the how
databases are created. This should then limit free form database
creations which could cause problems at scale. The solutions is to
add the model to the controller model list and when requesting to
access that model, that the uuid exists in that table.

Destroying a model will also remove the model from the controller model
list table. If the model fails to be removed from the db, but is removed
from state, then it will still be cleaned up. The model uuid will be
tainted and can not be used again. This should be fine as model uuid are
randomised.

The model manager domain was also partially implemented to allow for
this to happen. The feedback here was the result of the sprint discussion
around where the state should live.

We should do something about UUIDs in juju there are 6 packages for uuid
imported. We should pick one and use that. I'm currently using utils/uuid
but would rather use an external one, or anything not within the utils
package.