Force through tracked db in dummy provider
For some reason in the feature tests we require the dummy provider to
be more than a provider. We actually expect it to be also a jujud.
This is stupid.

Anyway, to get the feature tests running, I push through a trackedDB
through the many levels of turd so that they pass.