Allow building schema gen
This patch is broken off from #15539. Now we're actually pushing
the DB into the APIServer, it means that we need to make the schema
gen aware of sqlite.

The solution is to just set CGO_ENABLE=1 and then install the
libsqlite3-dev apt package where required. This shouldn't be required
for local development or for production builds, but is required for
testing. This isn't ideal, as we now have three different setups for
local, production and now testing. Once we move to snaps for the
controller agent, we can then at least merge local and production,
with testing fixed in the future when we change to a RESTful API.