Controller model table schema
This patch is broken off from #15539. This defines the model table
schema. This schema is intended to be small in scope. The current
sole responsibility of this table is to ensure that the model exists
and can therefore be opened by the db accessor worker.

Potential future work could include number of accesses, last usage,
but for now we purely just want to restrict open/view access. If the
value is within the table, consider it fair game for access.

Additionally, I've moved the schema outside of the database package
and into the domain package. This feels more like a logical location.