Model table/manager domain
This patch is broken off from #15539. The following implements the
model-list via the model manager domain. The purpose of this set
of changes to ensure that we restrict the how databases are created.
This should then limit free form database creations which could cause
problems at scale. The solutions is to add the model to the controller
model list and when requesting to access that model, that the uuid
exists in that table.

We should do something about UUIDs in juju there are 6 packages for uuid
imported. We should pick one and use that. I'm currently using utils/uuid
but would rather use an external one, or anything not within the utils
package.

Otherwise the code is rather simple, it just creates a service and state
layer which records the uuid into the controller model table. In the future
we'll need to change this so that it records the right model into the right
db, but until then we can just live with it like this.