Wire up model table domain to the facade
The following wires up the table domain into the facade. When creating
or destroying a model it should correctly insert the UUID into the
model table. Which we can then exploit for checking if the dbaccessor
can get the table or not.

The code is simple, push the creation through the modelmanager facade
and call the create and delete at the correct location.

Destroying a model will also remove the model from the controller model
list table. If the model fails to be removed from the db, but is removed
from state, then it will still be cleaned up. The model uuid will be
tainted and can not be used again. This should be fine as model uuid are
randomised.